Remove local file related APIs and dependencies to fs from client library for browsers; Update test codes after removing dependency to fs; Split common.js to common.core.js, common.node.js & common.browser.js;
