+++ browser/azure-storage.queue.export.js
@@ -39,7 +39,7 @@ AzureStorage.createQueueServiceWithSas = function(hostUri, sasToken) {
   return new QueueService(null, null, hostUri, sasToken);
 };
 
-var azureCommon = require('../lib/common/common');
+var azureCommon = require('../lib/common/common.browser');
 var StorageServiceClient = azureCommon.StorageServiceClient;
 var SharedKey = azureCommon.SharedKey;
