+++ gruntfile.js
@@ -44,7 +44,9 @@ module.exports = function(grunt) {
           'lib/common/filters/linearretrypolicyfilter.js',
           'lib/common/filters/exponentialretrypolicyfilter.js',
           'lib/common/services/storageutilities.js',
-          'lib/services/blob/blobservice.js',
+          'lib/services/blob/blobservice.core.js',
+          'lib/services/blob/blobservice.node.js',
+          'lib/services/blob/blobservice.browser.js',
           'lib/services/blob/models/blobresult.js',
           'lib/services/blob/models/containerresult.js',
           'lib/services/blob/models/leaseresult.js',
@@ -58,7 +60,9 @@ module.exports = function(grunt) {
           'lib/services/table/tablebatch.js',
           'lib/services/table/tablequery.js',
           'lib/services/table/tableutilities.js',
-          'lib/services/file/fileservice.js',
+          'lib/services/file/fileservice.core.js',
+          'lib/services/file/fileservice.node.js',
+          'lib/services/file/fileservice.browser.js',
           'lib/services/file/fileutilities.js',
           'lib/services/file/models/shareresult.js',
           'lib/services/file/models/directoryresult.js',
