+++ lib/azure-storage.js
@@ -78,7 +78,7 @@ exports.createTableServiceWithSas = function (hostUri, sasToken) {
  * Blob client exports.
  * @ignore
  */
-var BlobService = require('./services/blob/blobservice');
+var BlobService = require('./services/blob/blobservice.node');
 
 exports.BlobService = BlobService;
 exports.BlobUtilities = require('./services/blob/blobutilities');
@@ -125,7 +125,7 @@ exports.createBlobServiceAnonymous = function (host) {
  * File client exports.
  * @ignore
  */
-var FileService = require('./services/file/fileservice');
+var FileService = require('./services/file/fileservice.node');
 
 exports.FileService = FileService;
 exports.FileUtilities = require('./services/file/fileutilities');
@@ -199,8 +199,7 @@ exports.createQueueServiceWithSas = function(hostUri, sasToken) {
 * @ignore
 */
 
-
-var azureCommon = require('./common/common');
+var azureCommon = require('./common/common.node');
 var StorageServiceClient = azureCommon.StorageServiceClient;
 var SharedKey = azureCommon.SharedKey;
 /**
