+++ lib/common/common.core.js
@@ -53,9 +53,7 @@ exports.BatchOperation = require('./streams/batchoperation');
 exports.ChunkAllocator = require('./streams/chunkallocator');
 exports.ChunkStream = require('./streams/chunkstream');
 exports.ChunkStreamWithStream = require('./streams/chunkstreamwithstream');
-exports.FileReadStream = require('./streams/filereadstream');
 exports.SpeedSummary = require('./streams/speedsummary');
-exports.rfs = require('./streams/readablefs');
 
 // Utilities
 exports.Constants = require('./util/constants');
