+++ lib/common/md5-wrapper/md5.browser.js
