+++ lib/common/md5-wrapper/package.json
@@ -2,6 +2,6 @@
   "name": "azure-storage-md5-wrapper",
   "author": "Microsoft Corporation",
   "license": "Apache-2.0",
-  "main": "node.js",
-  "browser": "browser.js"
+  "main": "md5.node.js",
+  "browser": "md5.browser.js"
 }
