+++ lib/common/streams/rangestream.js
@@ -14,7 +14,7 @@
 // limitations under the License.
 // 
 
-var azureCommon = require('./../common');
+var azureCommon = require('./../common.core');
 var azureutil = azureCommon.util;
 var Constants = require('./../util/constants');
 var EventEmitter = require('events').EventEmitter;
