+++ lib/common/util/util.js
@@ -14,8 +14,6 @@
 // limitations under the License.
 // 
 
-var fs = require('fs');
-var path = require('path');
 var _ = require('underscore');
 var util = require('util');
 var url = require('url');
@@ -529,8 +527,6 @@ exports.isStreamPaused = function (object) {
   }
   return false;
 };
-
-exports.pathExistsSync = fs.existsSync ? fs.existsSync : path.existsSync;
 
 /**
 * Parse copy progress string in the format of bytesCopied/totalBytes
