+++ lib/services/blob/models/blobresult.js
@@ -17,7 +17,7 @@
 // Module dependencies.
 var _ = require('underscore');
 
-var azureCommon = require('./../../../common/common');
+var azureCommon = require('./../../../common/common.core');
 var azureutil = azureCommon.util;
 var Constants = azureCommon.Constants;
 var HeaderConstants = Constants.HeaderConstants;
