+++ lib/services/blob/models/blocklistresult.js
@@ -17,7 +17,7 @@
 // Module dependencies.
 var _ = require('underscore');
 
-var azureCommon = require('./../../../common/common');
+var azureCommon = require('./../../../common/common.core');
 var xmlbuilder = azureCommon.xmlbuilder;
 var Constants = azureCommon.Constants;
