+++ lib/services/file/fileservice.core.js
@@ -15,7 +15,6 @@
 // 
 
 // Module dependencies.
-var fs = require('fs');
 var qs = require('querystring');
 var url = require('url');
 var util = require('util');
@@ -23,7 +22,7 @@ var _ = require('underscore');
 var extend = require('extend');
 var path = require('path');
 
-var azureCommon = require('./../../common/common');
+var azureCommon = require('./../../common/common.core');
 var Md5Wrapper = require('./../../common/md5-wrapper');
 var azureutil = azureCommon.util;
 var SR = azureCommon.SR;
@@ -45,7 +44,6 @@ var SpeedSummary = azureCommon.SpeedSummary;
 var ChunkAllocator = azureCommon.ChunkAllocator;
 var ChunkStream = azureCommon.ChunkStream;
 var ChunkStreamWithStream = azureCommon.ChunkStreamWithStream;
-var FileReadStream = azureCommon.FileReadStream;
 var FileRangeStream = require('./internal/filerangestream');
 
 // Models requires
@@ -2268,87 +2266,6 @@ FileService.prototype.getFileToText = function (share, directory, file, optionsO
 };
 
 /**
-* Downloads an Azure file into a file.
-* (Not recommended for use in the JavaScript Client Library)
-*
-* @this {FileService}
-* @param {string}             share                                       The share name.
-* @param {string}             directory                                   The directory name. Use '' to refer to the base directory.
-* @param {string}             file                                        The file name. File names may not start or end with the delimiter '/'.
-* @param {string}             localFileName                               The local path to the file to be downloaded.
-* @param {object}             [options]                                   The request options.
-* @param {boolean}            [options.skipSizeCheck]                     Skip the size check to perform direct download.
-*                                                                         Set the option to true for small files.
-*                                                                         Parallel download and speed summary won't work with this option on.
-* @param {SpeedSummary}       [options.speedSummary]                      The download tracker objects.
-* @param {int}                [options.parallelOperationThreadCount]      Parallel operation thread count.
-* @param {string}             [options.rangeStart]                        Return only the bytes of the file in the specified range.
-* @param {string}             [options.rangeEnd]                          Return only the bytes of the file in the specified range.
-* @param {boolean}            [options.useTransactionalMD5]               When set to true, Calculate and send/validate content MD5 for transactions.
-* @param {boolean}            [options.disableContentMD5Validation]       When set to true, MD5 validation will be disabled when downloading files.
-* @param {LocationMode}       [options.locationMode]                      Specifies the location mode used to decide which location the request should be sent to. 
-*                                                                         Please see StorageUtilities.LocationMode for the possible values.
-* @param {int}                [options.timeoutIntervalInMs]               The timeout interval, in milliseconds, to use for the request.
-* @param {int}                [options.clientRequestTimeoutInMs]          The timeout of client requests, in milliseconds, to use for the request.
-* @param {int}                [options.maximumExecutionTimeInMs]          The maximum execution time, in milliseconds, across all potential retries, to use when making this request.
-*                                                                         The maximum execution time interval begins at the time that the client begins building the request. The maximum
-*                                                                         execution time is checked intermittently while performing requests, and before executing retries.
-* @param {string}             [options.clientRequestId]                   A string that represents the client request ID with a 1KB character limit.
-* @param {bool}               [options.useNagleAlgorithm]                 Determines whether the Nagle algorithm is used; true to use the Nagle algorithm; otherwise, false.
-*                                                                         The default value is false.
-* @param {errorOrResult}      callback                                    `error` will contain information if an error occurs; 
-*                                                                         otherwise `[result]{@link FileResult}` will contain the file information.
-*                                                                         `response` will contain information related to this operation.
-* @return {SpeedSummary}
-* 
-* @example
-* var azure = require('azure-storage');
-* var FileService = azure.createFileService();
-* FileService.getFileToLocalFile('taskshare', taskdirectory', 'task1', 'task1-download.txt', function(error, serverFile) {
-*   if(!error) {
-*     // file available in serverFile.file variable
-*   }
-*/
-FileService.prototype.getFileToLocalFile = function (share, directory, file, localFileName, optionsOrCallback, callback) {
-  var userOptions;
-  azureutil.normalizeArgs(optionsOrCallback, callback, function (o, c) { userOptions = o; callback = c; });
-  userOptions.speedSummary = userOptions.speedSummary || new SpeedSummary(file);
-
-  validate.validateArgs('getFileToLocalFile', function (v) {
-    v.string(share, 'share');
-    v.stringAllowEmpty(directory, 'directory');
-    v.string(file, 'file');
-    v.string(localFileName, 'localFileName');
-    v.shareNameIsValid(share);
-    v.callback(callback);
-  });
-
-  var options = extend(true, {}, userOptions);
-
-  var writeStream = fs.createWriteStream(localFileName);
-
-  this.getFileToStream(share, directory, file, writeStream, options, function (error, responseFile, response) {
-    if (error) {
-      // make sure writeStream is closed / destroyed to avoid locking issues
-      if (writeStream.close) {
-        writeStream.close();
-      }
-
-      if (azureutil.pathExistsSync(localFileName)) {
-        // If the download failed from the beginning, remove the file.
-        fs.unlink(localFileName, function () {
-          callback(error, responseFile, response);
-          return;
-        });
-      }
-    }
-    callback(error, responseFile, response); 
-  });
-  
-  return options.speedSummary;
-};
-
-/**
 * Provides a stream to read from a file.
 *
 * @this {FileService}
@@ -2771,78 +2688,6 @@ FileService.prototype.createFileFromText = function (share, directory, file, tex
       self._createRanges(share, directory, file, text, null, 0, length - 1, options, callback);
     }
   });
-};
-
-/**
-* Uploads a file to storage from a local file. If the file already exists on the service, it will be overwritten.
-* (Not recommended for use in the JavaScript Client Library)
-*
-* @this {FileService}
-* @param {string}             share                                         The share name.
-* @param {string}             directory                                     The directory name. Use '' to refer to the base directory.
-* @param {string}             file                                          The file name. File names may not start or end with the delimiter '/'.
-* @param (string)             localFileName                                 The local path to the file to be uploaded.
-* @param {object}             [options]                                     The request options.
-* @param {SpeedSummary}       [options.speedSummary]                        The download tracker objects;
-* @param {bool}               [options.storeFileContentMD5]                 Specifies whether the file's ContentMD5 header should be set on uploads. 
-*                                                                           The default value is false for files.
-* @param {bool}               [options.useTransactionalMD5]                 Calculate and send/validate content MD5 for transactions.
-* @param {object}             [options.contentSettings]                     The file's content settings.
-* @param {string}             [options.contentSettings.contentType]         The MIME content type of the file. The default type is application/octet-stream.
-* @param {string}             [options.contentSettings.contentEncoding]     The content encodings that have been applied to the file.
-* @param {string}             [options.contentSettings.contentLanguage]     The natural languages used by this resource.
-* @param {string}             [options.contentSettings.cacheControl]        The file service stores this value but does not use or modify it.
-* @param {string}             [options.contentSettings.contentDisposition]  The file's content disposition.
-* @param {string}             [options.contentSettings.contentMD5]          The file's MD5 hash.
-* @param {object}             [options.metadata]                            The metadata key/value pairs.
-* @param {LocationMode}       [options.locationMode]                        Specifies the location mode used to decide which location the request should be sent to. 
-*                                                                           Please see StorageUtilities.LocationMode for the possible values.
-* @param {int}                [options.timeoutIntervalInMs]                 The server timeout interval, in milliseconds, to use for the request.
-* @param {int}                [options.clientRequestTimeoutInMs]            The timeout of client requests, in milliseconds, to use for the request.
-* @param {int}                [options.maximumExecutionTimeInMs]            The maximum execution time, in milliseconds, across all potential retries, to use when making this request.
-*                                                                           The maximum execution time interval begins at the time that the client begins building the request. The maximum
-*                                                                           execution time is checked intermittently while performing requests, and before executing retries.
-* @param {string}             [options.clientRequestId]                     A string that represents the client request ID with a 1KB character limit.
-* @param {bool}               [options.useNagleAlgorithm]                   Determines whether the Nagle algorithm is used; true to use the Nagle algorithm; otherwise, false.
-*                                                                           The default value is false.
-* @param {errorOrResult}      callback                                      `error` will contain information if an error occurs; 
-*                                                                           otherwise `[result]{@link FileResult}` will contain the file information.
-*                                                                           `response` will contain information related to this operation.
-* @return {SpeedSummary}
-*/
-FileService.prototype.createFileFromLocalFile = function (share, directory, file, localFileName, optionsOrCallback, callback) {
-  var userOptions;
-  azureutil.normalizeArgs(optionsOrCallback, callback, function (o, c) { userOptions = o; callback = c; });
-
-  validate.validateArgs('createFileFromLocalFile', function (v) {
-    v.string(share, 'share');
-    v.stringAllowEmpty(directory, 'directory');
-    v.string(file, 'file');
-    v.string(localFileName, 'localFileName');
-    v.shareNameIsValid(share);
-    v.callback(callback);
-  });
-
-  var options = extend(true, {}, userOptions);
-  options.speedSummary = options.speedSummary || new SpeedSummary(file);
-
-  var self = this;
-  fs.stat(localFileName, function(error, stat) {
-    if (error) {
-      callback(error);
-    } else {
-      self.createFile(share, directory, file, stat.size, options, function(error) {
-        if(error) {
-          callback(error);
-        } else {
-          var stream = new FileReadStream(localFileName, {calcContentMd5: options.storeFileContentMD5});
-          self._createFileFromChunkStream(share, directory, file, stream, stat.size, options, callback);
-        }
-      });
-    }
-  });
-
-  return options.speedSummary;
 };
 
 /**
