+++ lib/services/file/models/directoryresult.js
@@ -15,7 +15,7 @@
 // 
 
 // Module dependencies.
-var HeaderConstants = require('./../../../common/common').Constants.HeaderConstants;
+var HeaderConstants = require('./../../../common/common.core').Constants.HeaderConstants;
 
 /**
 * Creates a new DirectoryResult object.
