+++ lib/services/file/models/shareresult.js
@@ -15,7 +15,7 @@
 // 
 
 // Module dependencies.
-var azureCommon = require('./../../../common/common');
+var azureCommon = require('./../../../common/common.core');
 var azureutil = azureCommon.util;
 var Constants = azureCommon.Constants;
 var HeaderConstants = Constants.HeaderConstants;
