+++ lib/services/queue/models/queuemessageresult.js
@@ -15,7 +15,7 @@
 // 
 
 // Module dependencies.
-var azureCommon = require('./../../../common/common');
+var azureCommon = require('./../../../common/common.core');
 var azureutil = azureCommon.util;
 var xmlbuilder = azureCommon.xmlbuilder;
 var Constants = azureCommon.Constants;
