+++ lib/services/queue/models/queueresult.js
@@ -15,7 +15,7 @@
 // 
 
 // Module dependencies.
-var Constants = require('./../../../common/common').Constants;
+var Constants = require('./../../../common/common.core').Constants;
 var HeaderConstants = Constants.HeaderConstants;
 
 /**
