+++ lib/services/queue/queueservice.js
@@ -19,7 +19,7 @@ var util = require('util');
 var _ = require('underscore');
 var extend = require('extend');
 
-var azureCommon = require('./../../common/common');
+var azureCommon = require('./../../common/common.core');
 var azureutil = azureCommon.util;
 var SR = azureCommon.SR;
 var validate = azureCommon.validate;
