+++ lib/services/table/internal/edmhandler.js
@@ -18,7 +18,7 @@ var _ = require('underscore');
 var util = require('util');
 var guid = require('node-uuid');
 
-var azureCommon = require('./../../../common/common');
+var azureCommon = require('./../../../common/common.core');
 var azureutil = azureCommon.util;
 var SR = azureCommon.SR;
