+++ lib/services/table/internal/odatahandler.js
@@ -17,7 +17,7 @@
 // Module dependencies.
 var util = require('util');
 
-var azureCommon = require('./../../../common/common');
+var azureCommon = require('./../../../common/common.core');
 var azureutil = azureCommon.util;
 var SR = azureCommon.SR;
 var Constants = azureCommon.Constants;
