+++ lib/services/table/internal/requesthandler.js
@@ -15,7 +15,7 @@
 // 
 
 var util = require('util');
-var azureCommon = require('./../../../common/common');
+var azureCommon = require('./../../../common/common.core');
 var WebResource = azureCommon.WebResource;
 var SR = azureCommon.SR;
 var Constants = azureCommon.Constants;
