+++ lib/services/table/internal/sharedkeytable.js
@@ -16,7 +16,7 @@
 
 // Module dependencies.
 var util = require('util');
-var azureCommon = require('./../../../common/common');
+var azureCommon = require('./../../../common/common.core');
 var SharedKey = azureCommon.SharedKey;
 var azureutil = azureCommon.util;
 var Constants = azureCommon.Constants;
