+++ lib/services/table/models/batchresult.js
@@ -15,7 +15,7 @@
 // 
 
 // Module dependencies.
-var azureCommon = require('./../../../common/common');
+var azureCommon = require('./../../../common/common.core');
 var Md5Wrapper = require('./../../../common/md5-wrapper');
 var StorageServiceClient = azureCommon.StorageServiceClient;
 var WebResource = azureCommon.WebResource;
