+++ lib/services/table/models/entityresult.js
@@ -15,7 +15,7 @@
 // 
 
 // Module dependencies.
-var azureCommon = require('./../../../common/common');
+var azureCommon = require('./../../../common/common.core');
 var Constants = azureCommon.Constants;
 var TableConstants = Constants.TableConstants;
 var HeaderConstants = Constants.HeaderConstants;
