+++ lib/services/table/tablebatch.js
@@ -17,7 +17,7 @@
 // Module dependencies.
 var extend = require('extend');
 
-var azureCommon = require('./../../common/common');
+var azureCommon = require('./../../common/common.core');
 var SR = azureCommon.SR;
 var validate = azureCommon.validate;
 var Constants = azureCommon.Constants;
