+++ lib/services/table/tablequery.js
@@ -18,7 +18,7 @@
 var _ = require('underscore');
 var util = require('util');
 
-var azureCommon = require('./../../common/common');
+var azureCommon = require('./../../common/common.core');
 var azureutil = azureCommon.util;
 var SR = azureCommon.SR;
 var QueryStringConstants =  azureCommon.Constants.QueryStringConstants;
