+++ lib/services/table/tableservice.js
@@ -19,7 +19,7 @@ var util = require('util');
 var extend = require('extend');
 var _ = require('underscore');
 
-var azureCommon = require('./../../common/common');
+var azureCommon = require('./../../common/common.core');
 var azureutil = azureCommon.util;
 var validate = azureCommon.validate;
 var SR = azureCommon.SR;
