+++ package.json
@@ -30,8 +30,7 @@
     "underscore": "~1.4.4",
     "validator": "~3.22.2",
     "xml2js": "0.2.7",
-    "xmlbuilder": "0.4.3",
-    "md5.js": "1.3.4"
+    "xmlbuilder": "0.4.3"
   },
   "devDependencies": {
     "browserify": "^13.3.0",
